Add back a piece removed years ago when cleaning
Someone accidentally removed this back in the changes after 0.16 and
it's been broken for some time. The variables pulled in by the extras
were never actually used in the binding and just ignored by the
method. This adds back the required piece to actually use the extras
being parsed by the eval_file_with_extras method as the name implies.